mentors-of-color-backend/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.34.95
botocore==1.34.95
certifi==2024.2.2
cffi==1.16.0
channels==3.0.5
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==3.0.2
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.10
django-allauth==0.61.1
django-auditlog==2.3.0
django-channels==0.7.0
django-cors-headers==4.3.1
django-filter==24.1
django-simple-history==3.4.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
gunicorn==22.0.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jmespath==1.0.1
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2024.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.9.0
urllib3==1.26.6
whitenoise==6.6.0
zope.interface==6.2
